+++ src/main/java/br/com/schumaker/force/framework/web/http/HttpVerb.java
@@ -1,4 +1,4 @@
-package br.com.schumaker.octopus.framework.web.http;
+package br.com.schumaker.force.framework.web.http;
 
 /**
  * The HttpVerb enum represents the various HTTP methods.
